requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
ollama==0.3.3
httpx==0.27.2
python-dotenv==1.0.1
rich==13.8.1
chromadb==0.5.11
selectolax==0.3.26
pytest==8.3.3
numpy==2.1.1
simpleaudio==1.0.4
pyttsx3==2.90
PyYAML==6.0.2

# API integrations
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
slack-sdk==3.27.1
spotipy==2.23.0
playwright==1.47.2
pyautogui==0.9.54
opencv-python==4.10.0.84
pytesseract==0.3.13
pyperclip==1.9.0
pygetwindow==0.0.9

# Voice and conversation dependencies
SpeechRecognition==3.10.0
whisper==1.1.10
sounddevice==0.4.6
scipy==1.11.4
